Rearrange some declarations for property testing the cluster
